Implementing sphere geometry

+++ pantaray/shading.hpp
@@ -26,8 +26,8 @@ namespace PantaRay {
         }
 
         Color Shade(Ray& ray, Intersection& intersection) {
-            int x = (int) float(intersection.u / size);
-            int y = (int) float(intersection.v / size);
+            int x = (int) double(intersection.u / size);
+            int y = (int) double(intersection.v / size);
 
             return (x + y) % 2 == 0 ? first : second;
         }
